+++ src/pages/basic-data/nvi/components/styles/nvi-admin-table-styles.ts
